Add files via upload
My first flutter application 
It's works with the esense flutter package and the esense earable.

https://www.esense.io/

This project was developed as part of the course Mobile Computing and Internet of things. It is a prototype, which combines the ear sensor with dancing. 
It is needed due to copyright issues that a new music asset is set for use the project.